[refactor/TK-138]: 패키지 구조 수종
