--- src/main/java/withbeetravel/dto/request/ChooseParticipantsRequest.java
@@ -1,17 +1,0 @@
-package withbeetravel.dto.request;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@Schema(description = "공동 결제 내역에 대한 정산 인원 리스트 DTO")
-public class ChooseParticipantsRequest {
-
-    @Schema(
-            description = "정산 인원 리스트",
-            example = "[17, 19, 22, 27]"
-    )
-    List<Long> travelMembersId;
-}
