+++ src/main/java/withbeetravel/dto/response/shardPayment/SharedPaymentRecordResponse.java
@@ -1,4 +1,4 @@
-package withbeetravel.dto.response;
+package withbeetravel.dto.response.shardPayment;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
