+++ src/main/java/withbeetravel/service/sharedPayment/SharedPaymentRecordService.java
@@ -1,7 +1,7 @@
 package withbeetravel.service.sharedPayment;
 
 import org.springframework.web.multipart.MultipartFile;
-import withbeetravel.dto.response.SharedPaymentRecordResponse;
+import withbeetravel.dto.response.shardPayment.SharedPaymentRecordResponse;
 import withbeetravel.dto.response.SuccessResponse;
 
 public interface SharedPaymentRecordService {
