+++ src/main/java/withbeetravel/service/travel/HoneyCapsuleService.java
@@ -1,6 +1,6 @@
 package withbeetravel.service.travel;
 
-import withbeetravel.dto.response.HoneyCapsuleResponse;
+import withbeetravel.dto.response.travel.HoneyCapsuleResponse;
 import withbeetravel.dto.response.SuccessResponse;
 
 import java.util.List;
